merge staging a0876a03 into master with GO from CQE

+++ Tensile/Source/client/include/ResultFileReporter.hpp
@@ -77,8 +77,8 @@ namespace Tensile
             std::string m_winnerSolution;
             int64_t     m_currSolutionIdx   = -1;
             int64_t     m_winnerSolutionIdx = -1;
-            int64_t     m_fastestGflops     = -1;
-            double      m_fasterTimeUS      = -1;
+            double      m_fastestGflops     = -1.0;
+            double      m_fasterTimeUS      = -1.0;
             // for merge rows
             int64_t                                                         m_currProbID = -1;
             std::map<int64_t, std::unordered_map<std::string, std::string>> m_probMap;
